Moved some non-templated CheckSumCombine functions into a source file to avoid duplicate definition issues.
