+++ msvc2015/Common/Common.vcxproj
@@ -223,6 +223,7 @@
     <ClCompile Include="..\..\network\MessageQueue.cpp" />
     <ClCompile Include="..\..\network\Networking.cpp" />
     <ClCompile Include="..\..\universe\Fighter.cpp" />
+    <ClCompile Include="..\..\util\CheckSums.cpp" />
     <ClCompile Include="..\..\util\EnumText.cpp" />
     <ClCompile Include="..\..\util\SaveGamePreviewUtils.cpp" />
     <ClCompile Include="..\..\util\ScopedTimer.cpp" />
