+++ util/CheckSums.h
@@ -1,6 +1,8 @@
 #ifndef _CheckSums_h_
 #define _CheckSums_h_
 
+#include "Export.h"
+
 #include <iostream>
 
 namespace CheckSums {
@@ -40,8 +42,7 @@ namespace CheckSums {
     }
 
     // fallback do nothing for unsupported types (without GetCheckSum functions)
-    void CheckSumCombine(...)
-    { std::cout << "CheckSumCombine(...)" << std::endl << std::endl; }
+    FO_COMMON_API void CheckSumCombine(...);
 
     // applies to pairs (including map value types)
     template <class C, class D>
@@ -75,33 +76,8 @@ namespace CheckSums {
         sum %= CHECKSUM_MODULUS;
     }
 
-    // doubles
-    void CheckSumCombine(unsigned int& sum, const double& t) {
-        std::cout << "CheckSumCombine(double): " << typeid(t).name() << std::endl << std::endl;
-        assert(DBL_MAX_10_EXP < 400);
-        if (t == 0.0)
-            return;
-        // biggest and smallest possible double should be ~10^(+/-308)
-        // taking log gives a number in the range +/- 309
-        // adding 400 gives numbers in the range ~0 to 800
-        // multiplying by 10'000 gives numbers in the range ~0 to 8'000'000
-        sum += static_cast<unsigned int>((std::log10(std::abs(t)) + 400.0) * 10000.0);
-        sum %= CHECKSUM_MODULUS;
-    }
-
-    // floats
-    void CheckSumCombine(unsigned int& sum, const float& t) {
-        std::cout << "CheckSumCombine(float): " << typeid(t).name() << std::endl << std::endl;
-        assert(FLT_MAX_10_EXP < 40);
-        if (t == 0.0f)
-            return;
-        // biggest and smallest possible float should be ~10^(+/-38)
-        // taking log gives a number in the range +/- 39
-        // adding 400 ives numbers in the range ~0 to 80
-        // multiplying by 100'000 gives numbers in the range ~0 to 8'000'000
-        sum += static_cast<unsigned int>((std::log10(std::abs(t)) + 40.0f) * 100000.0f);
-        sum %= CHECKSUM_MODULUS;
-    }
+    FO_COMMON_API void CheckSumCombine(unsigned int& sum, const double& t);
+    FO_COMMON_API void CheckSumCombine(unsigned int& sum, const float& t);
 }
 
 #endif // _CheckSums_h_
